[feat] 프로필 호버 시, 팝업 띄우도록 구현
